requirements.txt
Flask-Bootstrap==3.3.0.1
Flask-Login==0.2.11
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.8
WTForms==2.0.2
Werkzeug==0.10.1
itsdangerous==0.24
onetimepass==0.3.0
six==1.9.0
